[FIX]: some bugs
