+++ src/helpers/enums.ts
@@ -1,0 +1,14 @@
+/* eslint-disable no-unused-vars */
+
+export enum LoadingStatus {
+  IDLE = 'IDLE',
+  PENDING = 'PENDING',
+  FULLFILLED = 'FULLFILLED',
+  REJECTED = 'REJECTED'
+}
+
+export enum Gender {
+  MALE = 'MALE',
+  FEMALE = 'FEMALE',
+  OTHER = 'OTHER',
+}
