+++ src/helpers/strings.ts
@@ -1,18 +1,3 @@
-/* eslint-disable no-unused-vars */
-
-export enum LoadingStatus {
-  IDLE = 'IDLE',
-  PENDING = 'PENDING',
-  FULLFILLED = 'FULLFILLED',
-  REJECTED = 'REJECTED'
-}
-
-export enum Gender {
-  MALE = 'MALE',
-  FEMALE = 'FEMALE',
-  OTHER = 'OTHER',
-}
-
 export const ApiConst = {
   twitterUrl: 'https://twitter.com',
   privacyPolicyUrl: 'https://docs.example.org/extras/legal/privacy-policy',
