+++ src/hooks/index.ts
@@ -1,4 +1,5 @@
 
 
-export * from './customHooks';
+export {
 
+}
